inference/pyproject.toml
[project]
name = "inference"
description = ""
version = "0.0.1"
authors = [
    { name = "Alessandro Albanesi", email = "alessandro.albanesi@studio.unibo.it" }
]
dependencies = [
    "scikit-learn",
    "numpy",
    "pyaml",
    "pymongo",
    "requests"
]


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
